refactor: rename piecepoly to piecewise
