+++ src/pynurbs/core/piecewise.py
