+++ tests/core/test_spline_basis.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from pynurbs.core.knotvector import ImmutableKnotVector
-from pynurbs.core.piecepoly import PiecewisePolynomial
+from pynurbs.core.piecewise import PiecewisePolynomial
 from pynurbs.core.polynomial import Polynomial
 from pynurbs.core.spline_basis import ImmutableSplineBasis
